Reorganize the tools
